+++ src/main/java/com/sdl/selenium/web/form/ITextField.java
@@ -19,6 +19,6 @@ public interface ITextField extends IWebLocator, Editable {
      */
     String getValue();
 
-    <T extends ITextField> T setPlaceholder(final String value, SearchType...searchType);
+    <T extends ITextField> T setPlaceholder(final String value, SearchType...searchTypes);
 
 }